backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic-settings==2.5.0
supabase==2.10.0
PyJWT==2.9.0
cryptography>=42.0.0
anthropic==0.39.0
python-multipart==0.0.12
httpx==0.27.0
